Componentize the zoom button container
Separated the buttons to handle zoom functions into its own component. Trying to clean up the code base
